Move RID observers to NetRIDObserversResource
