+++ monitoring/uss_qualifier/resources/netrid/__init__.py
@@ -1,1 +1,2 @@
 from .service_providers import NetRIDServiceProviders
+from .observers import NetRIDObserversResource
