+++ monitoring/uss_qualifier/run_locally_rid.sh
@@ -50,17 +50,25 @@ echo '{
               }
             ]
           }
+        },
+        "netrid_observers": {
+          "resource_type": "netrid.NetRIDObserversResource",
+          "dependencies": {
+            "auth_adapter": "utm_auth"
+          },
+          "specification": {
+            "observers": [
+              {
+                "name": "uss2",
+                "observation_base_url": "http://host.docker.internal:8073/riddp/observation"
+              }
+            ]
+          }
         }
       }
     },
     "rid": {
-      "rid_version": "F3411-19",
-      "observers": [
-        {
-          "name": "uss2",
-          "observation_base_url": "http://host.docker.internal:8073/riddp/observation"
-        }
-      ]
+      "rid_version": "F3411-19"
     }
 }' > ${CONFIG_LOCATION}
